+++ drbd/drbd_receiver.c
@@ -1934,10 +1934,10 @@ STATIC void drbd_uuid_dump(struct drbd_conf *mdev, char *text, u64 *uuid)
 	}
 	INFO("%s %016llX:%016llX:%016llX:%016llX\n",
 	     text,
-	     uuid[Current],
-	     uuid[Bitmap],
-	     uuid[History_start],
-	     uuid[History_end]);
+	     (unsigned long long)uuid[Current],
+	     (unsigned long long)uuid[Bitmap],
+	     (unsigned long long)uuid[History_start],
+	     (unsigned long long)uuid[History_end]);
 }
 
 /*
